+++ src/Data/Work.js
@@ -1,0 +1,54 @@
+import React from "react";
+
+export const Apps = {
+  key: 0,
+  galleryIndex: 0,
+  width: 1582,
+  height: 1638,
+  captionRightEdge: 820,
+  image: "url(images/apps.png)",
+  href: "https://apps.apple.com/us/app/mollie/id1473455257?ls=1",
+  label: "Enjoy managing payments on mobile. ↗",
+  project: "Mollie Apps",
+  description:
+    "During the last quarter of 2019 I designed Mollie’s mobile apps to enable people to quickly manage payments and watch their business grow.",
+};
+export const MollieVideo = {
+  key: 1,
+  galleryIndex: 1,
+  width: 2880,
+  height: 1800,
+  captionRightEdge: 1600,
+  href: "videos/airpods-max.mp4",
+  poster: "images/airpods-max.jpg",
+  label: "Enjoy managing payments on mobile. ↗",
+  project: "Mollie Video",
+  description:
+    "During the last quarter of 2019 I designed Mollie’s mobile apps to enable people to quickly manage payments and watch their business grow.",
+};
+export const Checkout = {
+  key: 2,
+  galleryIndex: 2,
+  width: 1492,
+  height: 1022,
+  captionRightEdge: 820,
+  image: "url(images/checkout.png)",
+  href: "https://apps.apple.com/us/app/mollie/id1473455257?ls=1",
+  label: "Enjoy managing payments on mobile. ↗",
+  project: "Mollie Checkout",
+  description:
+    "During the first quarter of 2019 I redesigned and developed the Checkout to close the technology gap, meet customer’s expectations of swiftly paying for their order with confidence.",
+};
+export const ApplePay = {
+  key: 3,
+  galleryIndex: 3,
+  width: 2880,
+  height: 1800,
+  captionRightEdge: 820,
+  href: "videos/airpods-max.mp4",
+  poster: "images/airpods-max.jpg",
+  label: "Enjoy managing payments on mobile. ↗",
+  project: "Mollie Video",
+  description:
+    "During the last quarter of 2019 I designed Mollie’s mobile apps to enable people to quickly manage payments and watch their business grow.",
+};
