--- src/Mamba/Button.js
@@ -1,14 +1,0 @@
-import styled from "styled-components";
-
-export const Button = styled.button`
-  font: inherit;
-  letter-spacing: -0.023rem;
-  padding: 8px 24px;
-  border-radius: 32px;
-  text-align: center;
-  color: var(--white);
-  background-color: var(--red);
-  border: none;
-  margin: 0;
-  cursor: pointer;
-`;
