+++ src/Mamba/Buttons.js
@@ -1,0 +1,112 @@
+import styled from "styled-components";
+
+const imageHeight = 250;
+
+export const Button = styled.button`
+  font: inherit;
+  letter-spacing: -0.023rem;
+  padding: 8px 24px;
+  border-radius: 32px;
+  text-align: center;
+  color: var(--white);
+  background-color: var(--red);
+  border: none;
+  margin: 0;
+  cursor: pointer;
+`;
+
+export const CloseButton = styled.button`
+  position: absolute;
+  right: 16px;
+  top: 16px;
+  height: 36px;
+  width: 36px;
+  border-radius: 32px;
+  background-color: var(--secondaryFill);
+  border: none;
+  margin: 0;
+  padding: 0;
+  cursor: pointer;
+  display: grid;
+  place-items: center;
+  z-index: 1;
+  outline: none;
+  opacity: 0;
+
+  &.is-morphed {
+    opacity: 1;
+    transition: opacity 0.37s cubic-bezier(0.52, 0.16, 0.24, 1) 0.37s;
+  }
+`;
+
+export const PreviousButton = styled.button`
+  position: absolute;
+  top: calc(${imageHeight}px / 2 - 20px);
+  left: -20px;
+  height: 52px;
+  width: 52px;
+  border-radius: 32px;
+  background-color: var(--secondaryFill);
+  border: none;
+  margin: 0;
+  padding: 0;
+  cursor: pointer;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 1;
+  transition: 0.25s background-color linear, 0.25s opacity linear;
+  outline: none;
+
+  svg path {
+    color: var(--primaryLabelFill);
+    transition: stroke 0.25s linear;
+  }
+
+  &:hover,
+  &:focus,
+  &:active {
+    background-color: var(--tertiaryFill);
+    color: var(--secondaryLabelFill);
+
+    svg path {
+      stroke: var(--secondaryLabelFill);
+    }
+  }
+`;
+
+export const NextButton = styled.button`
+  position: absolute;
+  top: calc(${imageHeight}px / 2 - 20px);
+  right: -20px;
+  height: 52px;
+  width: 52px;
+  border-radius: 32px;
+  background-color: var(--secondaryFill);
+  border: none;
+  margin: 0;
+  padding: 0;
+  outline: none;
+  cursor: pointer;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 1;
+  transition: background-color 0.25s linear, opacity 0.25s linear;
+
+  path {
+    color: var(--primaryLabelFill);
+    transition: stroke 0.25s linear;
+  }
+
+  &:hover,
+  &:focus,
+  &:active {
+    background-color: var(--tertiaryFill);
+    color: var(--secondaryLabelFill);
+
+    path {
+      stroke: var(--secondaryLabelFill);
+    }
+  }
+`;
