--- src/Mamba/CloseButton.js
@@ -1,25 +1,0 @@
-import styled from "styled-components";
-
-export const CloseButton = styled.button`
-  position: absolute;
-  right: 16px;
-  top: 16px;
-  height: 36px;
-  width: 36px;
-  border-radius: 32px;
-  background-color: var(--secondaryFill);
-  border: none;
-  margin: 0;
-  padding: 0;
-  cursor: pointer;
-  display: grid;
-  place-items: center;
-  z-index: 1;
-  outline: none;
-  opacity: 0;
-
-  &.is-morphed {
-    opacity: 1;
-    transition: opacity 0.37s cubic-bezier(0.52, 0.16, 0.24, 1) 0.37s;
-  }
-`;
