+++ src/Mamba/MorphVideo.js
@@ -28,6 +28,7 @@ const MorphContainer = styled.li`
   }
 `;
 
+// No image needed for Video
 const MorphContent = styled.figure`
   position: absolute;
   top: 0;
@@ -44,7 +45,7 @@ const MorphContent = styled.figure`
     bottom: 10vh;
   }
 
-  @media (max-width: 1059px) {
+  @media (max-width: 1023px) {
     &.is-morphed {
       top: 16vh;
       bottom: 42vh;
@@ -117,7 +118,7 @@ const MorphOpenButton = styled.button`
     transition: opacity 0.129s cubic-bezier(0.52, 0.16, 0.24, 1);
   }
 
-  @media (max-width: 1059px) {
+  @media (max-width: 1023px) {
     &.is-morphed {
       transition: opacity 0.24s cubic-bezier(0.52, 0.16, 0.24, 1);
     }
@@ -138,6 +139,7 @@ const MorphCaption = styled(Caption)`
   }
 `;
 
+// Video added
 const MorphVideo = styled.video`
   width: 100%;
   display: block;
@@ -266,8 +268,148 @@ export function LargeMorphVideo(props) {
             style={{ color: "var(--red)" }}
             href={props.href}
           >
+            <strong>{props.project}. </strong>
+            {props.description}
+            <br />
             {props.label}
           </Link>
+        </FigCaption>
+      </MorphContent>
+
+      <MorphOpenButton
+        onClick={() => handleMorph(contentRef)}
+        className={isMorphed && "is-morphed"}
+      >
+        <MorphCaption className={isMorphed && "is-morphed"}>
+          <strong>{props.project}</strong>
+          <br />
+          Learn more
+        </MorphCaption>
+      </MorphOpenButton>
+    </MorphContainer>
+  );
+}
+
+export function SmallMorphVideo(props) {
+  const contentWidth = props.width;
+  const contentHeight = props.height;
+  const galleryIndex = props.galleryIndex;
+  const sendMorphstate = props.sendMorphstate;
+  const transformedIndex = props.galleryIndex - props.pageIndex;
+
+  // Distance between the center of the image and its optical right edge in the coordinate system of the native image resolution
+  const captionRightEdges = props.captionRightEdge;
+
+  const contentRef = useRef(null);
+  const captionRef = useRef(null);
+
+  const [viewportHeight, setViewportHeight] = useState(0);
+  const [viewportWidth, setViewportWidth] = useState(0);
+  const [isMorphed, setIsMorphed] = useState(false);
+  const [morphTop, setMorphTop] = useState(0);
+  const [morphLeft, setMorphLeft] = useState(0);
+  const [captionX, setCaptionX] = useState(0);
+
+  function layoutCaptions() {
+    const scale =
+      viewportWidth / viewportHeight > contentWidth / contentHeight
+        ? viewportHeight / contentHeight
+        : viewportWidth / contentWidth;
+    const xPos = viewportWidth / 2.0 + captionRightEdges * scale;
+    const x = Math.round(xPos);
+    setCaptionX(x);
+  }
+
+  function handleMorph(ref) {
+    setIsMorphed(!isMorphed);
+    sendMorphstate(!isMorphed);
+    const screenWidth = window.innerWidth;
+    const screenOffset = screenWidth - viewportWidth;
+    const galleryOffset = galleryIndex * screenOffset;
+    setMorphTop(-ref.current.getBoundingClientRect().top);
+    setMorphLeft(-ref.current.getBoundingClientRect().left - galleryOffset);
+  }
+
+  useEffect(() => {
+    contentRef.current, captionRef.current;
+  }, []);
+
+  useEffect(() => {
+    const myObserver = new ResizeObserver((entries) => {
+      entries.forEach((entry) => {
+        setViewportWidth(entry.contentRect.width);
+        setViewportHeight(entry.contentRect.height);
+      });
+    });
+
+    myObserver.observe(contentRef.current);
+  }, []);
+
+  useEffect(() => {
+    layoutCaptions();
+  }, [layoutCaptions]);
+
+  useEffect(() => {
+    const dissmissModal = () => {
+      setIsMorphed(false);
+    };
+    window.addEventListener("resize", dissmissModal);
+    return () => window.removeEventListener("resize", dissmissModal);
+  }, []);
+
+  useEffect(() => {
+    isMorphed
+      ? (document.body.style = "overflow: hidden")
+      : (document.body.style = `overflow: ""`);
+  }, [isMorphed]);
+
+  return (
+    <MorphContainer
+      style={{
+        top: isMorphed ? morphTop : null,
+        left: isMorphed ? morphLeft : null,
+        transform: `translate3d( ${100 * galleryIndex}%, 0, 0)`,
+        opacity: transformedIndex === 0 ? 1 : 0,
+        zIndex: isMorphed ? 20 : null,
+      }}
+      className={isMorphed && "is-morphed"}
+    >
+      <MorphCloseButton
+        onClick={() => (setIsMorphed(!isMorphed), sendMorphstate(!isMorphed))}
+        className={isMorphed && "is-morphed"}
+      >
+        <CloseIcon />
+      </MorphCloseButton>
+
+      <MorphContent ref={contentRef} className={isMorphed && "is-morphed"}>
+        <div
+          style={{
+            width: isMorphed ? "100%" : "calc(100% - 48px)",
+            borderRadius: 4,
+            margin: "auto",
+            border: "3px solid #111",
+            transition: MorphTransition,
+          }}
+        >
+          <MorphVideo
+            controls
+            preload='metadata'
+            poster={props.poster}
+            className={isMorphed && "is-morphed"}
+          >
+            <source src={props.href} />
+          </MorphVideo>
+        </div>
+
+        <FigCaption
+          ref={captionRef}
+          style={{
+            transform: `translate3d(${captionX}px, 0, 0)`,
+          }}
+          className={isMorphed && "is-morphed"}
+        >
+          <strong>{props.project}. </strong>
+          {props.description}
         </FigCaption>
       </MorphContent>
