+++ src/Mamba/index.js
@@ -1,7 +1,6 @@
 export * from "./Text";
 export * from "./Icon";
-export * from "./Button";
-export * from "./CloseButton";
+export * from "./Buttons";
 export * from "./Gallery";
 export * from "./Header";
 export * from "./MorphBox";
